backend/requirements.txt
alembic==1.17.2
bcrypt==5.0.0
click==8.3.1
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.6.0
python-dotenv==0.21.0
SQLAlchemy==1.4.0
typing_extensions==4.15.0
Werkzeug==2.2.2
pytest
pytest-flask
